abstract logo and shortcode into own component for reusability
