+++ src/routes/(admin)/organization/OrganizationForm.svelte
@@ -1,6 +1,4 @@
 <script lang="ts">
-	import Check from '@lucide/svelte/icons/check';
-	import ImageIcon from '@lucide/svelte/icons/image';
 	import FileText from '@lucide/svelte/icons/file-text';
 	import LinkIcon from '@lucide/svelte/icons/link';
 	import Settings from '@lucide/svelte/icons/settings';
@@ -8,12 +6,12 @@
 	import { Button } from '$lib/components/ui/button';
 	import * as Form from '$lib/components/ui/form/index.js';
 	import { Input } from '$lib/components/ui/input/index.js';
+	import LogoField from '$lib/components/LogoField.svelte';
+	import ShortcodeField from '$lib/components/ShortcodeField.svelte';
 	import { type Infer, superForm, fileProxy } from 'sveltekit-superforms';
 	import { zod4Client } from 'sveltekit-superforms/adapters';
 	import type { PageData } from './$types';
-	import { page } from '$app/stores';
 	import { resolve } from '$app/paths';
-	import { browser } from '$app/environment';
 	import { invalidateAll } from '$app/navigation';
 	import { editOrganizationSchema, type EditOrganizationSchema } from './schema';
 	import { toast } from 'svelte-sonner';
@@ -36,13 +34,16 @@
 		currentPlatformGuideName = data?.platformGuideFilename ?? null;
 	});
 
+	let logoInput: HTMLInputElement | undefined = $state();
+	let pdfInput: HTMLInputElement;
+
 	const form = superForm(orgData || data.form, {
 		validators: zod4Client(editOrganizationSchema),
 		dataType: 'form',
 		onResult: ({ result }) => {
-			if (fileInput) {
-				fileInput.value = '';
-				fileInput.dispatchEvent(new Event('change', { bubbles: true }));
+			if (logoInput) {
+				logoInput.value = '';
+				logoInput.dispatchEvent(new Event('change', { bubbles: true }));
 			}
 			if (pdfInput) {
 				pdfInput.value = '';
@@ -55,44 +56,10 @@
 	});
 
 	const { form: formData, enhance, tainted, submitting } = form;
-	const logoFile = fileProxy(form, 'logo');
 	const platformGuideFile = fileProxy(form, 'platform_guide');
-
-	let fileInput: HTMLInputElement;
-	let pdfInput: HTMLInputElement;
-
-	const selectedFileName = $derived($logoFile?.[0]?.name || null);
-	let previewUrl = $state<string | null>(null);
-	const logoDisplayName = $derived(
-		selectedFileName ?? (currentLogoUrl ? currentLogoUrl.split('/').pop() : null)
-	);
 
 	const selectedPdfName = $derived($platformGuideFile?.[0]?.name || null);
 	const pdfDisplayName = $derived(selectedPdfName ?? currentPlatformGuideName);
-
-	$effect(() => {
-		const file = $logoFile?.[0];
-		if (!browser || !file) {
-			previewUrl = null;
-			return;
-		}
-		const url = URL.createObjectURL(file);
-		previewUrl = url;
-		return () => {
-			URL.revokeObjectURL(url);
-		};
-	});
-
-	let copied = $state(false);
-
-	function copyPortalUrl() {
-		const fullUrl = `${$page.url.origin}/${$formData.shortcode ?? ''}`;
-		navigator.clipboard.writeText(fullUrl);
-		copied = true;
-		setTimeout(() => {
-			copied = false;
-		}, 2000);
-	}
 
 	async function deleteLogo() {
 		const res = await fetch('/api/organization/logo', {
@@ -170,109 +137,9 @@
 						<Form.FieldErrors />
 					</Form.Field>
 
-					<Form.Field {form} name="shortcode" class="flex flex-col gap-1.5">
-						<Form.Control>
-							{#snippet children({ props })}
-								<Form.Label class="font-semibold">Shortcode</Form.Label>
-								<div
-									class="border-input ring-offset-background bg-card focus-within:border-ring focus-within:ring-ring/50 flex h-9 items-center rounded-md border shadow-xs transition-[color,box-shadow] focus-within:ring-[3px]"
-								>
-									<span class="text-muted-foreground pl-3 text-sm whitespace-nowrap">
-										{$page.url.host}/
-									</span>
-									<input
-										{...props}
-										class="text-foreground min-w-0 flex-1 bg-transparent py-1 pr-3 text-sm outline-none"
-										bind:value={$formData.shortcode}
-									/>
-									<button
-										type="button"
-										onclick={copyPortalUrl}
-										class="text-primary hover:text-primary/80 flex items-center gap-1 px-3 text-sm font-medium whitespace-nowrap"
-									>
-										{#if copied}
-											<Check class="h-4 w-4" /> Copied
-										{:else}
-											Copy Portal URL
-										{/if}
-									</button>
-								</div>
-							{/snippet}
-						</Form.Control>
-						<Form.FieldErrors />
-					</Form.Field>
+					<ShortcodeField {form} />
 
-					<Form.Field {form} name="logo" class="flex flex-col gap-1.5">
-						<Form.Control>
-							{#snippet children({ props })}
-								<Form.Label class="font-semibold">Logo</Form.Label>
-								<input
-									{...props}
-									type="file"
-									accept="image/png,image/jpeg,image/webp"
-									bind:files={$logoFile}
-									bind:this={fileInput}
-									hidden
-								/>
-								<div class="flex items-center gap-3">
-									{#if previewUrl}
-										<img
-											src={previewUrl}
-											alt="Logo preview"
-											class="border-border bg-muted h-10 w-10 shrink-0 rounded-md border object-contain"
-										/>
-									{:else if currentLogoUrl}
-										<img
-											src={currentLogoUrl}
-											alt="Current logo"
-											class="border-border bg-muted h-10 w-10 shrink-0 rounded-md border object-contain"
-										/>
-									{:else}
-										<div
-											class="border-border bg-muted flex h-10 w-10 shrink-0 items-center justify-center rounded-md border"
-										>
-											<ImageIcon class="text-muted-foreground h-4 w-4" />
-										</div>
-									{/if}
-
-									<div
-										class="border-input bg-card flex h-9 flex-1 items-center rounded-md border shadow-xs"
-									>
-										<div class="flex min-w-0 flex-1 items-center gap-2 px-3">
-											<ImageIcon class="text-muted-foreground h-4 w-4 shrink-0" />
-											<span
-												class={[
-													'truncate text-sm',
-													logoDisplayName ? 'text-foreground' : 'text-muted-foreground'
-												]}
-											>
-												{logoDisplayName ?? 'No file selected'}
-											</span>
-										</div>
-										<button
-											type="button"
-											onclick={() => fileInput.click()}
-											class="text-primary hover:text-primary/80 px-3 text-sm font-medium"
-										>
-											Change
-										</button>
-										{#if currentLogoUrl && !selectedFileName}
-											<button
-												type="button"
-												onclick={deleteLogo}
-												class="text-muted-foreground hover:text-destructive pr-3"
-												aria-label="Delete logo"
-											>
-												<Trash_2 class="h-4 w-4" />
-											</button>
-										{/if}
-									</div>
-								</div>
-								<p class="text-muted-foreground text-xs">PNG, JPG, WebP, max 2MB</p>
-							{/snippet}
-						</Form.Control>
-						<Form.FieldErrors />
-					</Form.Field>
+					<LogoField {form} {currentLogoUrl} onDelete={deleteLogo} bind:fileInputRef={logoInput} />
 
 					<Form.Field {form} name="platform_guide" class="flex flex-col gap-1.5">
 						<Form.Control>
